better camera
tweaked camera settings to make x-axis movement more responsive, and dialed back the y-axis sensitivity. Also shortened the distance between the 3 rigs so movement along the y-axis feels less dramatic.

Also, added code to movement script that so that sappy looks in the direction you are moving with the camera